Added a comment
As a coder for the first time, I found no way/instruction to check a task, which made me quite anxious. So, I wrote a comment on how to check a task for a first-time coder to feel free and pull requests for the benefit of open source. Thank you.